requirements.txt
fastmcp
python-dotenv
requests

